LayoutTests/accessibility/input-auto-fill-button.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body>
<input type="password" id="password">
<script>
description("This tests that there exists an accessibility element for the AutoFill button.");

if (!window.internals || !window.accessibilityController)
    testFailed("You must run this test in the test tool and have implemented window.internals and window.accessibilityController.");
else {
    internals.setShowAutoFillButton(document.getElementById("password"), true);

    shouldBeEqualToString('accessibilityController.accessibleElementById("password").childAtIndex(0).description', "AXDescription: autofill");
    shouldBeEqualToString('accessibilityController.accessibleElementById("password").childAtIndex(0).role', "AXRole: AXButton");
}
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
