LayoutTests/accessibility/input-auto-fill-button-expected.txt

This tests that there exists an accessibility element for the AutoFill button.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.accessibleElementById("password").childAtIndex(0).description is "AXDescription: autofill"
PASS accessibilityController.accessibleElementById("password").childAtIndex(0).role is "AXRole: AXButton"
PASS successfullyParsed is true

TEST COMPLETE
